Split out some tests
